+++ src/test/java/com/exercise/basic/string/Bai7Test.java
@@ -1,8 +1,8 @@
-package com.exercise.basic;
+package com.exercise.basic.string;
 
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class Bai7Test {
